Implement stricter matching of number literals.
Before this commit we did not enforce equality of number literals.
So a query like `$s = 0;` did also match on `size = 1;`.
With this commit, integer literals are compared based on their numerical
value so `$s = 0;` would match on `size = 0;` but also `size = 0x0;` or
`0b0`. Floats (3.14) now require an exact string match.