Add support for the "strict:" label.
strict can be used to disable some of weggli's greedier matching features.
For now, statement unwrapping is disabled so you can search for function calls,
that do not check their return value. In addition, identifier matching is
more strict so searching for foo() won't match on a->foo() or std::foo().