cli.rs: give --extensions example
The current CLI `--help` text does not state what is the format of the `--extensions` or `-e` flags, so this commit makes it more verbose by giving an explicit example within the help text.

This will help prevent users to get confused when they use Weggli like this:
```
$ weggli -e '.h' -X "..." <path>
No files to parse. Exiting...
$ weggli -e '*.h' -X "..." <path>
No files to parse. Exiting...
```